+++ VisuaLigue/src/main/java/ca/ulaval/glo2004/visualigue/ui/models/FrameModel.java
@@ -1,14 +1,14 @@
 package ca.ulaval.glo2004.visualigue.ui.models;
 
 import ca.ulaval.glo2004.visualigue.ui.models.actors.ActorModel;
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.LongProperty;
+import javafx.beans.property.SimpleLongProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableMap;
 
 public class FrameModel extends ModelBase {
 
-    public IntegerProperty time = new SimpleIntegerProperty();
+    public LongProperty time = new SimpleLongProperty();
     public ObservableMap<String, ActorModel> actorModels = FXCollections.observableHashMap();
 
 }
